+++ home.html
@@ -6,6 +6,7 @@
     <title>home</title>
     <link rel="shortcut icon" href="togo.png" type="image/x-icon">
     <script src="https://cdn.tailwindcss.com"></script>
+    <script type="module" src="home.js"></script>
 </head>
 <body class="flex flex-col min-h-screen bg-[#ff534f]">
     <div class="min-w-[100vw] flex items-center mx-auto pr-4 bg-[#ffd75b]">
