+++ login.html
@@ -5,13 +5,13 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>login</title>
     <link rel="shortcut icon" href="togo.png" type="image/x-icon">
-    <script type="module" src="auth.js"></script>
+    <script src="login.js" type="module"></script>
     <script src="https://cdn.tailwindcss.com"></script>
 </head>
 <body class="flex min-h-screen items-center justify-center bg-[#ff534f]">
     <div class="bg-[#ffd75b] text-[4vw] md:text-[2rem] p-10 rounded-xl flex flex-col gap-2">
         <div class="text-[5vw] md:text-[3rem] mb-2">login</div>
-        <input class="bg-[#f9ac40] rounded-lg py-1 px-2.5 placeholder-black"  id="username" type="text" placeholder="username">
+        <input class="bg-[#f9ac40] rounded-lg py-1 px-2.5 placeholder-black"  id="email" type="text" placeholder="email">
         <input class="bg-[#f9ac40] rounded-lg py-1 px-2.5 placeholder-black" id="password" type="password" placeholder="password">
         <button class="mt-2 bg-[#ff534f] px-3 py-2 rounded-xl" id="signup">login</button>
         <div id="display"></div>
