+++ signup.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>signup</title>
     <link rel="shortcut icon" href="togo.png" type="image/x-icon">
-    <script type="module" src="auth.js"></script>
+    <script src="signup.js" type="module"></script>
     <script src="https://cdn.tailwindcss.com"></script>
 </head>
 <body class="flex min-h-screen items-center justify-center bg-[#ff534f]">
